added user authentication.

+++ graphql/resolvers/booking.js
@@ -2,7 +2,10 @@ import Booking from '../../models/booking'
 import { singleEvent, user } from './merge'
 
 export default {
-  bookings: async () => {
+  bookings: async (args, req) => {
+    if (!req.isAuth) {
+      throw new Error('UnAuthenticated!')
+    }
     try {
       let bookings = await Booking.find().lean() // add limit if U want..
       return bookings.map((booking) => {
@@ -17,17 +20,20 @@ export default {
     }
   },
 
-  bookEvent: async ({ eventId }) => {
+  bookEvent: async ({ eventId }, req) => {
+    if (!req.isAuth) {
+      throw new Error('UnAuthenticated!')
+    }
     try {
       const userAlreadyHasBooking = await Booking.findOne({
-        user: '5eff7bef39b1874d91ccb10d',
+        user: req.userId,
         event: eventId
       }).lean()
 
       if (userAlreadyHasBooking) throw new Error('User already has this booking!')
 
       const booking = await new Booking({
-        user: '5eff7bef39b1874d91ccb10d',
+        user: req.userId,
         event: eventId
       }).save()
 
@@ -40,7 +46,11 @@ export default {
       throw error
     }
   },
-  cancelBooking: async ({ bookingId }) => {
+  cancelBooking: async ({ bookingId }, req) => {
+    if (!req.isAuth) {
+      throw new Error('UnAuthenticated!')
+    }
+
     const booking = await Booking.findById(bookingId).populate('event').lean()
     if (booking) {
       await Booking.deleteOne({ _id: bookingId })
